Moved all PageData generics functions inwards
This avoids the dart type errors when building a
widget T with an argument A (T Funcion(A)) when using that
function outside of the defined class (PageData<T extends Widget,A>).
The issue occures when downcasting a class (i.e. in a List<UntypedPageData>,
where we loos T and A)
and then accessing a builder method that still relies strictly on T and A.
Then dart crashes because it can not resolve the original PageData<T,A>

This also  required to add a pageRoute builder in order to not expose
any building methods to the outside.

Added a widget test that excercises this behavior, in order to avoid
a silent bug creep.